Serving/fastapi_onnx/requirements.txt
fastapi
uvicorn[standard]
onnxruntime-gpu   # provides TensorRT & CUDA EPs automatically
transformers 